Load data about shifts from server

+++ domain/src/main/kotlin/com.nstu.technician.domain/model/document/Contractor.kt
@@ -3,7 +3,7 @@ package com.nstu.technician.domain.model.document
 import com.nstu.technician.domain.model.common.Address
 
 class Contractor(
-    val oid: Int,
+    val oid: Long,
     val name: String,
     val address: Address,
     val INN: String
